Fix some code
